refactor and fix tests
