+++ lychee-lib/src/client.rs
@@ -22,19 +22,17 @@ use http::{
 use log::{debug, warn};
 use octocrab::Octocrab;
 use regex::RegexSet;
-use reqwest::{header, redirect, Url};
+use reqwest::{header, redirect};
 use reqwest_cookie_store::CookieStoreMutex;
 use secrecy::{ExposeSecret, SecretString};
 use typed_builder::TypedBuilder;
 
 use crate::{
-    chain::{Chain, ClientRequestChains, RequestChain},
+    chain::RequestChain,
     checker::file::FileChecker,
     checker::{mail::MailChecker, website::WebsiteChecker},
     filter::{Excludes, Filter, Includes},
-    quirks::Quirks,
     remap::Remaps,
-    types::uri::github::GithubUri,
     utils::fragment_checker::FragmentChecker,
     Base, BasicAuthCredentials, ErrorKind, Request, Response, Result, Status, Uri,
 };
@@ -385,24 +383,28 @@ impl ClientBuilder {
             include_mail: self.include_mail,
         };
 
-        Ok(Client {
+        let website_checker = WebsiteChecker::new(
+            self.method,
+            self.retry_wait_time,
+            self.max_retries,
             reqwest_client,
+            self.accepted,
             github_client,
+            self.require_https,
+            self.plugin_request_chain,
+        );
+
+        Ok(Client {
             remaps: self.remaps,
             filter,
-            max_retries: self.max_retries,
-            retry_wait_time: self.retry_wait_time,
-            method: self.method,
-            accepted: self.accepted,
-            require_https: self.require_https,
-            fragment_checker: FragmentChecker::new(),
             email_checker: MailChecker::new(),
+            website_checker,
             file_checker: FileChecker::new(
                 self.base,
                 self.fallback_extensions,
                 self.include_fragments,
             ),
-            plugin_request_chain: self.plugin_request_chain,
+            fragment_checker: FragmentChecker::new(),
         })
     }
 }
@@ -413,48 +415,23 @@ impl ClientBuilder {
 /// options.
 #[derive(Debug, Clone)]
 pub struct Client {
-    /// Underlying `reqwest` client instance that handles the HTTP requests.
-    reqwest_client: reqwest::Client,
-
-    /// Optional GitHub client that handles communications with GitHub.
-    github_client: Option<Octocrab>,
-
     /// Optional remapping rules for URIs matching pattern.
     remaps: Option<Remaps>,
 
     /// Rules to decided whether each link should be checked or ignored.
     filter: Filter,
 
-    /// Maximum number of retries per request before returning an error.
-    max_retries: u64,
+    /// A checker for website URLs.
+    website_checker: WebsiteChecker,
 
-    /// Initial wait time between retries of failed requests. This doubles after
-    /// each failure.
-    retry_wait_time: Duration,
+    /// A checker for file URLs.
+    file_checker: FileChecker,
 
-    /// HTTP method used for requests, e.g. `GET` or `HEAD`.
-    ///
-    /// The same method will be used for all links.
-    method: reqwest::Method,
-
-    /// Set of accepted return codes / status codes.
-    ///
-    /// Unmatched return codes/ status codes are deemed as errors.
-    accepted: Option<HashSet<StatusCode>>,
-
-    /// Requires using HTTPS when it's available.
-    ///
-    /// This would treat unencrypted links as errors when HTTPS is available.
-    require_https: bool,
+    /// A checker for email URLs.
+    email_checker: MailChecker,
 
     /// Caches Fragments
     fragment_checker: FragmentChecker,
-
-    plugin_request_chain: RequestChain,
-
-    file_checker: FileChecker,
-
-    email_checker: MailChecker,
 }
 
 impl Client {
@@ -545,132 +522,7 @@ impl Client {
         uri: &Uri,
         credentials: Option<BasicAuthCredentials>,
     ) -> Result<Status> {
-        let default_chain: RequestChain = Chain::new(vec![
-            Box::<Quirks>::default(),
-            Box::new(credentials),
-            Box::new(WebsiteChecker::new(
-                self.retry_wait_time,
-                self.max_retries,
-                self.reqwest_client.clone(),
-                self.accepted.clone(),
-            )),
-        ]);
-
-        match self.check_website_inner(uri, &default_chain).await {
-            Status::Ok(code) if self.require_https && uri.scheme() == "http" => {
-                if self
-                    .check_website_inner(&uri.to_https()?, &default_chain)
-                    .await
-                    .is_success()
-                {
-                    Ok(Status::Error(ErrorKind::InsecureURL(uri.to_https()?)))
-                } else {
-                    // HTTPS is not available for this URI,
-                    // so the original HTTP URL is fine.
-                    Ok(Status::Ok(code))
-                }
-            }
-            s => Ok(s),
-        }
-    }
-
-    /// Checks the given URI of a website.
-    ///
-    /// Unsupported schemes will be ignored
-    ///
-    /// Note: we use `inner` to improve compile times by avoiding monomorphization
-    ///
-    /// # Errors
-    ///
-    /// This returns an `Err` if
-    /// - The URI is invalid.
-    /// - The request failed.
-    /// - The response status code is not accepted.
-    pub async fn check_website_inner(&self, uri: &Uri, default_chain: &RequestChain) -> Status {
-        // Workaround for upstream reqwest panic
-        if validate_url(&uri.url) {
-            if matches!(uri.scheme(), "http" | "https") {
-                // This is a truly invalid URI with a known scheme.
-                // If we pass that to reqwest it would panic.
-                return Status::Error(ErrorKind::InvalidURI(uri.clone()));
-            }
-            // This is merely a URI with a scheme that is not supported by
-            // reqwest yet. It would be safe to pass that to reqwest and it
-            // wouldn't panic, but it's also unnecessary, because it would
-            // simply return an error.
-            return Status::Unsupported(ErrorKind::InvalidURI(uri.clone()));
-        }
-
-        let request = self
-            .reqwest_client
-            .request(self.method.clone(), uri.as_str())
-            .build();
-
-        let request = match request {
-            Ok(r) => r,
-            Err(e) => return e.into(),
-        };
-
-        let status = ClientRequestChains::new(vec![&self.plugin_request_chain, default_chain])
-            .traverse(request)
-            .await;
-
-        self.handle_github(status, uri).await
-    }
-
-    // Pull out the heavy machinery in case of a failed normal request.
-    // This could be a GitHub URL and we ran into the rate limiter.
-    // TODO: We should try to parse the URI as GitHub URI first (Lucius, Jan 2023)
-    async fn handle_github(&self, status: Status, uri: &Uri) -> Status {
-        if status.is_success() {
-            return status;
-        }
-
-        if let Ok(github_uri) = GithubUri::try_from(uri) {
-            let status = self.check_github(github_uri).await;
-            // Only return GitHub status in case of success
-            // Otherwise return the original error, which has more information
-            if status.is_success() {
-                return status;
-            }
-        }
-
-        status
-    }
-
-    /// Check a `uri` hosted on `GitHub` via the GitHub API.
-    ///
-    /// # Caveats
-    ///
-    /// Files inside private repositories won't get checked and instead would
-    /// be reported as valid if the repository itself is reachable through the
-    /// API.
-    ///
-    /// A better approach would be to download the file through the API or
-    /// clone the repo, but we chose the pragmatic approach.
-    async fn check_github(&self, uri: GithubUri) -> Status {
-        let Some(client) = &self.github_client else {
-            return ErrorKind::MissingGitHubToken.into();
-        };
-        let repo = match client.repos(&uri.owner, &uri.repo).get().await {
-            Ok(repo) => repo,
-            Err(e) => return ErrorKind::GithubRequest(e).into(),
-        };
-        if let Some(true) = repo.private {
-            // The private repo exists. Assume a given endpoint exists as well
-            // (e.g. `issues` in `github.com/org/private/issues`). This is not
-            // always the case but simplifies the check.
-            return Status::Ok(StatusCode::OK);
-        } else if let Some(endpoint) = uri.endpoint {
-            // The URI returned a non-200 status code from a normal request and
-            // now we find that this public repo is reachable through the API,
-            // so that must mean the full URI (which includes the additional
-            // endpoint) must be invalid.
-            return ErrorKind::InvalidGithubUrl(format!("{}/{}/{endpoint}", uri.owner, uri.repo))
-                .into();
-        }
-        // Found public repo without endpoint
-        Status::Ok(StatusCode::OK)
+        self.website_checker.check_website(uri, credentials).await
     }
 
     /// Checks a `mailto` URI.
@@ -689,16 +541,6 @@ impl Client {
             }
         }
     }
-}
-
-// Check if the given `Url` would cause `reqwest` to panic.
-// This is a workaround for https://github.com/lycheeverse/lychee/issues/539
-// and can be removed once https://github.com/seanmonstar/reqwest/pull/1399
-// got merged.
-// It is exactly the same check that reqwest runs internally, but unfortunately
-// it `unwrap`s (and panics!) instead of returning an error, which we could handle.
-fn validate_url(url: &Url) -> bool {
-    http::Uri::try_from(url.as_str()).is_err()
 }
 
 /// A shorthand function to check a single URI.
@@ -740,7 +582,7 @@ mod tests {
         chain::{ChainResult, Handler, RequestChain},
         mock_server,
         test_utils::get_mock_client_response,
-        Request, Status, Uri,
+        ErrorKind, Request, Status, Uri,
     };
 
     #[tokio::test]
@@ -989,9 +831,14 @@ mod tests {
     #[tokio::test]
     async fn test_avoid_reqwest_panic() {
         let client = ClientBuilder::builder().build().client().unwrap();
-        // This request will fail, but it won't panic
+        // This request will result in an Unsupported status, but it won't panic
         let res = client.check("http://\"").await.unwrap();
-        assert!(res.status().is_error());
+
+        assert!(matches!(
+            res.status(),
+            Status::Unsupported(ErrorKind::BuildRequestClient(_))
+        ));
+        assert!(res.status().is_unsupported());
     }
 
     #[tokio::test]
